fix(260414-smt): store increment_seconds as Float to preserve chess.com fractional increments
SmallInteger rounded '10+0.1' to increment_seconds=0, losing the fractional
bonus that chess.com emits for some bullet games. Changed the column and
Python types to Float so parse_base_and_increment preserves the raw value.

Dev DB migration rolled back and re-applied to pick up the corrected type.

+++ app/models/game.py
@@ -78,9 +78,9 @@ class Game(Base):
     base_time_seconds: Mapped[int | None] = mapped_column(
         SmallInteger, nullable=True
     )  # starting clock in seconds
-    increment_seconds: Mapped[int | None] = mapped_column(
-        SmallInteger, nullable=True
-    )  # increment per move in seconds
+    increment_seconds: Mapped[float | None] = mapped_column(
+        Float, nullable=True
+    )  # increment per move in seconds (Float: chess.com emits fractional values like 0.1s)
 
     # Flags
     rated: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
